+++ src/plugin.cpp
@@ -18,7 +18,8 @@ void init(rack::Plugin* p) {
 	p->addModel(modelEightFace);
 	p->addModel(modelEightFaceX2);
 	p->addModel(modelMidiCat);
-	p->addModel(modelMidiCatEx);
+	p->addModel(modelMidiCatMem);
+	p->addModel(modelMidiCatMap);
 	p->addModel(modelSipo);
 	p->addModel(modelFourRounds);
 	p->addModel(modelArena);